# Conflicts:
#	client/src/main/java/com/alibaba/nacos/client/config/NacosConfigService.java
#	client/src/main/java/com/alibaba/nacos/client/config/http/HttpAgent.java
#	client/src/main/java/com/alibaba/nacos/client/config/http/MetricsHttpAgent.java
#	client/src/main/java/com/alibaba/nacos/client/config/impl/ClientWorker.java
#	config/src/main/java/com/alibaba/nacos/config/server/service/LongPollingService.java

+++ client/src/main/java/com/alibaba/nacos/client/naming/beat/BeatReactor.java
@@ -165,7 +165,7 @@ public class BeatReactor implements Closeable {
             long nextTime = beatInfo.getPeriod();
             try {
                 JsonNode result = serverProxy.sendBeat(beatInfo, BeatReactor.this.lightBeatEnabled);
-                long interval = result.get("clientBeatInterval").asInt();
+                long interval = result.get("clientBeatInterval").asLong();
                 boolean lightBeatEnabled = false;
                 if (result.has(CommonParams.LIGHT_BEAT_ENABLED)) {
                     lightBeatEnabled = result.get(CommonParams.LIGHT_BEAT_ENABLED).asBoolean();
